api/src/requirements.txt
numpy
pandas
flask-restful
mysql-connector-python

